half()

+++ src/inference.py
@@ -7,7 +7,8 @@ from PIL import Image
 import os
 import math
 import time
-from .models import TextEncoder
+from models import TextEncoder
+# from .models import TextEncoder
 from transformers import AutoTokenizer
 try:
     from torchvision.transforms import InterpolationMode
@@ -111,7 +112,7 @@ class Tester():
         img = self._preprocess(img)
         with no_grad():
             if isHalf:
-                image_embedding = self.ImageEncoder.encode_image(img.to(self.device)) # Output : N x 512
+                image_embedding = self.ImageEncoder.encode_image(img.to(self.device)).half() # Output : N x 512
             else:
                 image_embedding = self.ImageEncoder.encode_image(img.to(self.device)).float() # Output : N x 512
